requirements.txt
asyncio
fastapi
pandas
python-decouple
starlette
uvicorn[standard]
uvloop